backend:refactor: split friendRequest tests
